More coherent WhiteListSanitizer behavior.
When allowed_attributes is set and allowed_tags is not set, all <script> and <form> tags are pruned. This means that the same tags are pruned as when neither allowed_attributes nor allowed_tags are set.

However if both allowed_attributes and allowed_tags are set, the behavior concerning <script> and <tag> forms depends on whether they are in the allowed_tags array or not. E.g. <form> tags can be untouched by the sanitizer if "form" is in the allowed_tags array. Note.- currently if allowed_tags is set, any tags not present in the array are stripped, not pruned; this is, their contents are present after sanitization. This is not fully coherent with the behavior when neither allowed_tags nor allowed_attributes are set; in this case <script> and <form> tags are pruned, not stripped, so their contents are removed as well.

HTML comments are removed by the sanitizer in any case, regardless of other options.