Fix failing unit test
One of the changes in released in Loofah 2.9.0 broke this test. I can't
tell if this was an intended or unintended side effect of the change to
support functions in shorthand CSS rule.

I'm not sure how well documented this change was, but Loofah's tests
were updated to reflect a different formatting in the commit that causes
this behavior. I think it's safe to say that stripping the whitespace
from rules like the one in this test is the expected behavior.

The test should be updated to reflect the new formatting.

Relevant links from Loofah:
https://github.com/flavorjones/loofah/releases/tag/v2.9.0
https://github.com/flavorjones/loofah/pull/200